requirements.txt
face_recognition>=1.3.0
numpy>=1.19.5
opencv-python>=4.5.3
Pillow>=8.3.1
scikit-learn>=0.24.2
pandas>=1.3.0
PyYAML>=6.0
matplotlib>=3.4.3
openpyxl>=3.0.9 